refactor(server): split the code in several files
